% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIDEOR.R
\docType{data}
\name{full_database}
\alias{full_database}
\title{Descriptive data title}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{pinyinnum}{}
\item{pinyintone}{}
\item{pinyinnone}{}
\item{traditional}{}
\item{simplified}{}
\item{MC}{}
\item{OC}{}
\item{radsup}{}
\item{morph}{}
\item{variant}{}
\item{Kroll}{}
\item{zdic}{}
\item{note}{}
}}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{CHIDEOR}
}
