% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIDEOR.R
\docType{package}
\name{CHIDEOR}
\alias{CHIDEOR}
\alias{CHIDEOR-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
CHIDEOR
A data package for CHIDEOR.

Use \code{data(package='CHIDEOR')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
NA
\link{CHIDEOR}
}
